requirements.txt
absl-py==2.1.0
amqp==5.2.0
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.7.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cmake==3.29.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
deepface==0.0.90
dlib @ file:///C:/Python/311/Scripts/dlib-19.24.1-cp311-cp311-win_amd64.whl#sha256=6f1a5ee167975d7952b28e0ce4495f1d9a77644761cf5720fb66d7c6188ae496
face-recognition==1.3.0
face-recognition-models==0.3.0
facenet-pytorch==2.6.0
filelock==3.13.4
fire==0.6.0
Flask==3.0.3
Flask-Cors==4.0.0
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
gast==0.5.4
gdown==5.1.0
google-pasta==0.2.0
grpcio==1.62.2
gunicorn==22.0.0
h5py==3.11.0
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.1.2
jax==0.4.26
Jinja2==3.1.3
keras==3.2.1
kiwisolver==1.4.5
kombu==5.3.7
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mediapipe==0.10.11
mkl==2021.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
mtcnn==0.1.1
mysql==0.0.3
mysql-connector-python==8.4.0
mysqlclient==2.2.4
namex==0.0.8
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
py-cpuinfo==9.0.0
pycparser==2.22
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
retina-face==0.0.17
rich==13.7.1
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
sounddevice==0.4.6
soupsieve==2.5
sympy==1.12
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tf_keras==2.16.0
thop==0.1.1.post2209072238
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.11
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.2
wrapt==1.16.0
